Add files via upload
This is the most recent version of our application, there are too many files so I have uploaded it as a .zip file. 
To make sure this app runs properly, you must install Node JS and the appropriate modules that can be found in the package.json file. An SQL server is also required, when you set up your server, make sure to change values like host, username, password, etc. in the connector.js file.